[0822]功能新增：动态出库调度策略（最优出库料箱选择算法，订单动态优先级算法，料箱命中多订单算法）, 出库传送带模型

+++ openTCS-Common/src/main/java/org/opentcs/util/persistence/v002/LocationTO.java
@@ -137,7 +137,7 @@ public class LocationTO
     @XmlType(propOrder = {"skuID", "quantity"})
     public static class SkuTO{
       private String skuID = "";
-      private Integer quantity = 0;
+      private Double quantity = 0.0;
       
       @XmlAttribute(required = true)
       public String getSkuID(){
@@ -150,10 +150,10 @@ public class LocationTO
       }
 
       @XmlAttribute(required = true)
-      public Integer getQuantity(){
+      public Double getQuantity(){
         return quantity;
       }
-      public SkuTO setQuantity(@Nonnull Integer quantity){
+      public SkuTO setQuantity(@Nonnull Double quantity){
         requireNonNull(quantity,"quantity");
         this.quantity = quantity;
         return this;
